# docs/devicemanager_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Device-Selection API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/selectables": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "deprecated; finds devices, device-groups and/or imports that match all provided filter-criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "selectables",
                    "deprecated"
                ],
                "summary": "deprecated selectables",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "result should include matching device-groups",
                        "name": "include_groups",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "result should include matching imports",
                        "name": "include_imports",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of local device ids; result devices must be in this list (if one is given)",
                        "name": "local_devices",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "adds full import-type and import path options to the result. device services are already complete, the name is a legacy artefact",
                        "name": "complete_services",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of protocol ids, that should be ignored",
                        "name": "filter_protocols",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "interaction that is not allowed in the result",
                        "name": "filter_interaction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "json encoded criteria list (model.FilterCriteriaAndSet like [{\u0026quot;function_id\u0026quot;:\u0026quot;\u0026quot;,\u0026quot;aspect_ids\u0026quot;:[],\u0026quot;device_class_id\u0026quot;:\u0026quot;\u0026quot;}])",
                        "name": "json",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "alternative to json; base64 encoded json of criteria list",
                        "name": "base64",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "alternative to json and base64 if only one filter criteria is needed",
                        "name": "function_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "alternative to json and base64 if only one filter criteria is needed",
                        "name": "device_class_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "alternative to json and base64 if only one filter criteria is needed; deprecated: alias for a single element aspect_ids",
                        "name": "aspect_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "alternative to json and base64 if only one filter criteria is needed; comma seperated list of aspect ids, all of which the same content variable has to carry",
                        "name": "aspect_ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.Selectable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v2/query/selectables": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "finds devices, device-groups and/or imports that match all provided filter-criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "selectables"
                ],
                "summary": "selectables",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "result should include matching devices",
                        "name": "include_devices",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "result should include matching device-groups",
                        "name": "include_groups",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "result should include matching imports",
                        "name": "include_imports",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "result should include all valid device id modifications",
                        "name": "include_id_modified",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "trim first element of import paths",
                        "name": "import_path_trim_first_element",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of device ids; result devices must be in this list (if one is given)",
                        "name": "devices",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of local device ids; result devices must be in this list (if one is given)",
                        "name": "local_devices",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "used in combination with local_devices to identify devices, default is the requesting user",
                        "name": "local_device_owner",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of attribute keys; result devices have these attributes (if one is given)",
                        "name": "filter_devices_by_attr_keys",
                        "in": "query"
                    },
                    {
                        "description": "criteria list",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/devicemodel.FilterCriteria"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.Selectable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "devicemodel.FilterCriteria": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "description": "deprecated: alias for a single element AspectIds",
                    "type": "string"
                },
                "aspect_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "device_class_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "interaction": {
                    "type": "string"
                }
            }
        },
...
        "github_com_SENERGY-Platform_device-selection_v2_pkg_model.ImportType": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImportTypeConfig"
                    }
                },
                "cost": {
                    "type": "integer"
                },
                "default_restart": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "output": {
                    "$ref": "#/definitions/models.ImportContentVariable"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
...
        "model.DeviceGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Import": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImportConfig"
                    }
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "import_type_id": {
                    "type": "string"
                },
                "kafka_topic": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "restart": {
                    "type": "boolean"
                }
            }
        },
        "model.PathOption": {
            "type": "object",
            "properties": {
                "aspectNode": {
                    "description": "deprecated: alias for a single element AspectNodes; holds the node with the alphabetically first id",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.AspectNode"
                        }
                    ]
                },
                "aspectNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AspectNode"
                    }
                },
                "characteristicId": {
                    "type": "string"
                },
                "configurables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Configurable"
                    }
                },
                "functionId": {
                    "type": "string"
                },
                "interaction": {
                    "$ref": "#/definitions/models.Interaction"
                },
                "isVoid": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                },
                "value": {}
            }
        },
        "model.PermSearchDevice": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attribute"
                    }
                },
                "creator": {
                    "type": "string"
                },
                "device_type_id": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "local_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "permissions": {
                    "$ref": "#/definitions/model.Permissions"
                },
                "shared": {
                    "type": "boolean"
                }
            }
        },
        "model.Permissions": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "boolean"
                },
                "r": {
                    "type": "boolean"
                },
                "w": {
                    "type": "boolean"
                },
                "x": {
                    "type": "boolean"
                }
            }
        },
        "model.Selectable": {
            "type": "object",
            "properties": {
                "device": {
                    "$ref": "#/definitions/model.PermSearchDevice"
                },
                "device_group": {
                    "$ref": "#/definitions/model.DeviceGroup"
                },
                "import": {
                    "$ref": "#/definitions/model.Import"
                },
                "importType": {
                    "$ref": "#/definitions/github_com_SENERGY-Platform_device-selection_v2_pkg_model.ImportType"
                },
                "servicePathOptions": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/model.PathOption"
                        }
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/devicemodel.Service"
                    }
                }
            }
        },
...
        "models.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "models.ImportConfig": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "models.Interaction": {
            "type": "string",
            "enum": [
                "event",
                "request",
                "event+request"
            ],
            "x-enum-varnames": [
                "EVENT",
                "REQUEST",
                "EVENT_AND_REQUEST"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}